main: resolve_url: parameterise http header language
Make it possible to set the language sent as the value of the Accept-Language
http header when curling web pages. This causes things to be localised to the
language selected, if the site serves data conditionally on this header.

This was previously added, but the language was not yet parameterisable.