http: use rust string processing rather than regex
The parse_contents function still fails on a lot of stuff due to attempting to
unwrap the result of the regex matching, which quite often is None.

Now, implement simple pure rust string processing logic to a) coerce multiple
line strings into one line, and b) strip leading and trailing whitespace,
rather than attempting to do these two things in a single regex.

It appears that there are no longer any dangerous unwraps in the resolve_url or
parse_contents functions.

This also means the Regex dependency can be dropped; it also seems to be the
case that the new implementation is lighter and faster.

Update tests with a particular failure case, which now passes.