src: main: update irc crate to 0.13.6
Update the irc crate, and update to the new IrcReactor API.

Add comments here and there to make understanding the code slightly easier.
This all could probably be thought out a bit better, and refactored at some
point since it started out as a quick hack and has had stuff added to it since.