Match Content-Type in response and do things accordingly
Do slightly different things depending on MIME types matched from the
Content-Type response header.

The first is to download different amounts of the file, particularly
downloading a maximum of 10MB for all images, to help with metadata extraction.
Otherwise, limit to 100KB as before.

for images and unknown types, which fail the HTML title parse, respond with a
new string of the textual MIME type and the file size, based on the
Content-Length response header.

This allows for responses such as text/plain or application/pdf for example.

If HTML parsing fails, produce and error as before.