Tidy up configuration loading
Tidy up loading code, making it an impl for the configuration struct, and load
all options including the irc options using serde. Remove the need to load the
config twice, once for the bot general options, and once again for the irc
handler.

This should also pave the way for potentially configuring a single instance of
the bot to connect to multiple servers, and for the possibility of writing back
to the config.

In the case of multiple instances, it still remains a problem how to specify
what database would be used for each connection, as it is unlikely to be a sane
default to use a single database for all connections.