Use standard paths for config file and database, make history optional
Use standard paths to find configuration/database files, provided by the
`directories` crate.

On linux, this is according to the the XDG spec [1], for example:

    ~/.config/url-bot-rs/config.toml

If $XDG_CONFIG_HOME is set this would be:

    $XDG_CONFIG_HOME/url-bot-rs/config.toml

Similarly, the database default path is now:

    ~/.local/share/url-bot-rs/history.db

This patch also makes history an optional feature, which can be specified
either by giving a path on the command line with `--conf` or `--db`, or via the
configuration.

As part of this change, various tunable parameters are tidied up and given
their own section in the configuration file, and a new section in the config is
added for database settings. This includes such settings as the user-agent
request header entry.

Resolves #19
Resolves #72
Closes #41 (phew!)

[1] https://standards.freedesktop.org/basedir-spec/basedir-spec-latest.html