Create application directories, and default config, if they don't exist
Create required directories and files automatically if they are missing. If the
configuration file doesn't exist, write a default-valued configuration file to
disk.

This is done automatically, and without prompting, due to concerns that the bot
is likely to be run daemonised, and required user intervention in this context
would be undesirable.

This effectively also adds the free feature of being able to generate the
example configuration data programmatically, so use this in the future to
implement a test to ensure that the example configuration matches the generated
default configuration data, and should therefore be relatively complete,
up-to-date, and correct.