Workaround for #79
Use a regular expression to check whether a URL contains unsafe characters,
according to RFC1738 [1]. If it does, skip the token.

This works around an upstream bug in `reqwest`, which may be fixed in due
course, but since a URL is invalid if it contains unsafe characters, this seems
like a valid check to include here.

Fixes #79.

[1] https://www.ietf.org/rfc/rfc1738.txt