Add coverage to travis build, and badge to README
This now spawns another job for getting test coverage data from the cargo build
using `cargo-tarpaulin` [1] [2], posting results and retrieving a badge from
codecov.io, added to the README.

Travis-wise this is a separate job, for which the overall run is allowed to
fast finish, since coverage is never going to fail the travis run, and also
takes a significantly longer than the other jobs, mainly to build the cargo
plugin. As such the coverage is allowed to continue and update the coverage
report, but in its own time.

[1] https://github.com/xd009642/tarpaulin
[2] https://crates.io/crates/cargo-tarpaulin