Build a custom user agent string, and make user agent non-configurable
Use "Mozilla/5.0 url-bot-rs/$PKG_VERSION" as a custom user-agent string.
This avoids difficulties with some servers behaving in certain ways when a
custom user agent isn't used, for example serving up gzipped content even when
accept-encoding doesn't specify this behaviour. This also ensures that the
current version of url-bot is sent with the request, rather than having it
hard-coded in the configuration.

Remove user-agent from the configuration, since it has turned out to be an
important factor in successfully retrieving titles in some cases (Amazon), and
doesn't really add any benefit in being user-configurable.