Follow redirections manually and provide cookies
This adds basic cookie support, and some sites requiring that cookies
are returned in repeat requests before they will deliver content now
work.

Redirection is carried out manually, by telling request to not follow
redirects, so that processing required to store and respond with cookies
during a redirect can be done.

Such sites provide a 3** HTTP response indicating a redirection, with a
new URL provided in the `location` HTTP response header giving the
redirection location. If the response has 1 or more `set-cookie`
headers, the client must store these. When the client then attempts to
retrieve the redirected URL, it must provide the stored cookies as a
semi-colon separated list of vey=value pairs as the value of the
`cookie` response header.

Examples are:

 - <https://www.tandfonline.com/doi/abs/10.1080/09637486.2018.1542666>
 - <https://www.sainsburys.co.uk/webapp/wcs/stores/servlet/gb/groceries/sainsburys-british-mature-grated-cheddar-500g>