Use html5ever to parse (and unescape) page titles
Use a real HTML parsing library, which should give better resilience than a
simple regex.

parse_title now obtains a DOM for the HTML and recurses it until we find a
title tag, then returns the contents of this tag.

This also side-steps issues seen with the `htmlescape` crate which is also now
unmaintained.