Updates to queries
Include a couple of changes of behaviour in handling queries. Since
noone else sees queries, these should be a private thing between
url-bot-rs and the user, so it isn't important that links have been
posted in queries, and when a link is posted in a channel, this should
not register as a pre-post. It is a feasible use case to use a query
with url-bot-rs to test links before posting in a channel, for example.
Furthermore, don't send notices in queries, since this isn't
particularly relevant or desirable.

- Respond with private message rather than notices in queries, even if
  set to reply with notice in channels.
- Query the database (if enabled) for previous postings, report as
  normal, but don't add posted links to the database.

Closes #105
Closes #93