Add support for fetching partial urls with no scheme
Now URLs which are posted without http:// or https:// will cause an
attempt to be made to retrieve a title for them.

For example, "rustup.rs", "google.co.uk" or "www.rust-lang.org" would
previously have been ignored, but now they will produce a title
response.

Add a static hashmap of top level domains, obtained from the build
output of the tld crate. Use this to identify tokens which look like
they contain a URL, if they contain a recognised TLD. If this is the
case, then prepend "http://" to the token and proceed as normal.

The TLD map is included statically since the perfect hash function used,
and the retrieval of the current list takes a significant time to
compile, when compiling the tld crate itself. Instead the build output
is used direcly and the compilation time of url-bot-rs is much shorter.
It should not be essential that the list is rigorously kept up to date.

Closes #132