Add invite, kick, and autosaving of the configuration
Now if a user for example posts `/invite url-bot-rs` in a channel, url-bot will
join that channel. This is enabled with the feature `invite`.

Additionally, enabling the `autosave` feature will result in the channel being
added to the list of channels in the configuration, which the bot will join on
startup. The same is true for `/kick`, which will remove the channel from the
list in the configuration.

Closes #62