Do not resolve titles for email addresses
<dmc> "hmm, https://user@mail.example.com is a valid URL indeed. An interesting side-effect :)"

Due to recent changes which resolve, e.g. "google.com" to
"http://google.com", email addresses are now processed, e.g.:

<dmc>  test@gmail.com
<u>  ⤷ Gmail

This is because the short url code attempts to parse the token as a url,
and, if failing, adds "http://" to the start of it, so an email address
now looks like a url with a username.

Now, reject any resulting urls which parse as having a username.

This shouldn't affect anyone who wants to submit urls with credentials,
they will simply have to use the full form of the url, including the
scheme, e.g.:

https://bob:mmhcheese@flailing-ninjas.com/

Thanks @polyzen!