Use localhost for default configuration
Rather than connecting to freenode with a default configuration, connect
to a server at localhost instead, which is presumably less problematic a
postition to start with, and aids local testing, since installing an
IRCd with a system package manager will most likely start a server on
localhost without any extra configuration.

As such the default configuration (the one written out by url-bot-rs if
given a non-existing file path does quite well) may be used as the basis
of local debugging.