Always join status channels when posting
Previously if the bot was kicked from a status channel, the list of
joined channels maintained by the IRC client isn't updated, so url-bot
will not attempt to rejoin the status channel it was kicked from as it
thinks it's already joined.

Now just always join, which is OK, since it's pretty much a no-op as far
as the IRC protocol is concerned.