Fix calculation of redirection location for relative locations
Fixing tests.

Many sites return e.g. "/path" as the redirection location, it seems
frequently for URLs with subdomains, since it's perhaps easier to let
the client work out the relative path from the path it already knows,
while from the server side it might not be so easy to tell.

Joining of "/path/" to even "http://base-url.org/otherpath/" is correctly
handled by the Url crate, to result in "http://base-url.org/path/".