Fix repost test: don't add to db if pre-post found, but later negated
Even if a pre-post is found, various configurations can cause the result
of this look up to be ignored, meaning another pre-post entry will be
added to the database.

Add a flag to preserve this state, even if the actual result of the
pre-post look up is masked by configuration.

Also fix 0.0.0.0 to 127.0.0.1 after rebase.