Use bundled sqlite for all test runs, except release
This means test runs are more resilient to network issues on the CI, and
the ubuntu repositories being inaccessible, for example, and leading to
misleading failures.

Release still uses a system packaged sqlite3-dev, since the release deb
package will depend on this, so building with statically linked sqlite
would not be the correct thing to do.