Tests: change port from 8084 to 28382
Change the port 8084, which seems to have issues (HTTP error 400) on
Github Ubuntu runner. Choose an arbitrary high range port, separated
from ports used elsewhere in the tests.

Some day it might be nice to have a method to find an appropriate port
automatically, or to overhaul the HTTP server used in testing, or
implement integration tests based on tokio or something along those
lines.

Currently the tests are a bit fragile, although, that said they have
done reasonably well so far.