Reinstate non-hardened systemd unit as default
The hardened systemd unit causes issues on Debian, so reinstate the old
non-hardened version as the default version used for Debian packaging,
but keep the hardened version in tree.

This way releases can continue (with Debian package artifact), until the
issues can be worked out with hardening on Debian.

The failing log on starting the unit on Debian is as follows:

    Apr 29 09:14:18 fv-az133-393 systemd[1]: Started url-bot-rs.
    Apr 29 09:14:18 fv-az133-393 url-bot-rs[13217]: INFO - directory `/home/url-bot-rs/.config/url-bot-rs` doesn't exist, creating it
    Apr 29 09:14:18 fv-az133-393 url-bot-rs[13217]: ERROR - Permission denied (os error 13)
    Apr 29 09:14:18 fv-az133-393 systemd[1]: url-bot-rs.service: Main process exited, code=exited, status=1/FAILURE
    Apr 29 09:14:18 fv-az133-393 systemd[1]: url-bot-rs.service: Failed with result 'exit-code'.
    Apr 29 09:14:29 fv-az133-393 systemd[1]: url-bot-rs.service: Scheduled restart job, restart counter is at 6.
    Apr 29 09:14:29 fv-az133-393 systemd[1]: Stopped url-bot-rs.